Show other dissemination category properly when it errors
We need the field to be wrapped in a div, which necessitates some
changes to the styling, because Rails puts fields with an error
inside a div causing html errors otherwise. The js was also hiding
the field *after* checking on page load whether it should be shown,
so you didn't see it if there was an error.