Allow multiple countries to be selected for studies
This stores multiple countries as a comma-separated string
in the DB, but overrides the setter/getter so that we can
deal with them as arrays in Ruby. Also modifies the country
helper so that we can get a list of country objects or a
string representing all the names for printing.