Add fields to ERB Status to control display in timeline
Because admins can manage the ERB status options, when we want to
display some nice text for each status change and set an icon colour,
we need that info to be in the db too. This adds a description field
(for the timeline entry) and an enum to say whether it's a positive,
negative or neutral change (for the icon colour).