Add a StudyInvite model and invite_token field to Users
The intention is that invites can be created, inviting a user to a particular
study. The model also records who did the inviting. Separately, a user can be
given an invite token, which we'll use to log them in without a password. The
user will then be allowed to edit anything they've been invited to.