Improve highlighting of flagged studies
After user feedback and testing, we realised that *lots* of studies
are likely to be flagged, and for significant amounts of time, so
colouring them all in bright red probably wasn't as helpful as it
seemed initially. This changes the list views such that only the
status badge gets the colouring, and adds a middle level, orange,
between red and green to make things more subtle.