Allow only one output type to be recorded at a time
Entering text into any of the textareas or text inputs inside the
outputs recording form automatically disables all but the currently
active form section.

Clearing the entered text (or, additionally, in the case of the "other
impacts" section, using the checkboxes to hide a previously edited
textarea) resets the disabled sections, allowing you to change your
mind and switch to recording a different output type.